requirement.txt
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
dnspython==2.6.1
ecdsa==0.18.0
email-validator==2.1.1
exceptiongroup==1.2.0
fastapi==0.110.0
h11==0.14.0
idna==3.6
jwt==1.3.1
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycparser==2.21
pydantic==2.6.4
pydantic-core==2.16.3
PyJWT==2.8.0
python-jose==3.3.0
python-multipart==0.0.9
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
typing-extensions==4.10.0
uvicorn==0.29.0
